# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Marketplace YarmarOK 1.0.0",
    "description": "Web platform where vendors can come together to sell their products or services to customers.\n\nThis is Marketplace YarmarOK Server based on the OpenAPI 3.0 specification\n\nSome useful links:\n- [The YarmarokMarketplace repository](https://github.com/YarmarokMarketplace/marketplace)",
    "contact": {
      "name": "API developers",
      "email": "estasjuk@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8081/api",
      "description": "Development Server"
    },
    {
      "url": "https://yarmarok.onrender.com/api",
      "description": "Production Server"
    }
  ],
  "consumes": [
    "application/json",
    "multipart/form-data"
  ],
  "produces": [
    "application/json",
    "multipart/form-data"
  ],
  "tags": [
    {
      "name": "Notices",
      "description": "Notices endpoints"
    }
  ],
  "paths": {
    "/notices": {
      "get": {
        "summary": "Get all notices",
        "tags": [
          "Notices"
        ],
        "description": "Find all existing notices",
        "operationId": "getAllNotices",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListOfNoticesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "There any notices for this user"
          }
        }
      },
      "post": {
        "summary": "Add new notice",
        "tags": [
          "Notices"
        ],
        "description": "Add new notice",
        "operationId": "addNotice",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/AddNoticeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddNoticeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "409": {
            "description": "Email already in use"
          }
        }
      }
    },
    "/notices/{category}": {
      "get": {
        "summary": "Get notices by choosen category",
        "tags": [
          "Notices"
        ],
        "description": "Find and show all existing notices by choosen category",
        "operationId": "getNoticesByCategory",
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "example": "auto",
            "schema": {
              "type": "string",
              "description": "Category for search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListOfNoticesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Any notices for the given category"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ListOfNoticesResponse": {
        "type": "object",
        "properties": {
          "result": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "totalResults": {
                  "type": "integer",
                  "example": 8
                },
                "totalPages": {
                  "type": "integer",
                  "example": 2
                },
                "page": {
                  "type": "integer",
                  "example": 1
                },
                "limit": {
                  "type": "integer",
                  "example": 6
                },
                "notices": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AddNoticeResponse"
                  }
                }
              }
            },
            "description": "List of notices found by reques"
          }
        }
      },
      "AddNoticeRequest": {
        "type": "object",
        "required": [
          "title",
          "description",
          "price",
          "location",
          "category",
          "photo"
        ],
        "properties": {
          "title": {
            "type": "string",
            "example": "The chair for sale"
          },
          "description": {
            "type": "string",
            "example": "Very good chair for your home"
          },
          "category": {
            "type": "string",
            "enum": [
              "auto",
              "business-and-services",
              "for-free",
              "children's-world",
              "home-and-garden",
              "help",
              "electricity",
              "spare-parts-for-transport",
              "fashion-and-style",
              "realty",
              "exchange",
              "repair",
              "work",
              "animals",
              "goods-to-win",
              "hobbies-recreation-sports"
            ],
            "example": "home-and-garden"
          },
          "location": {
            "type": "string",
            "example": "Kyiv"
          },
          "price": {
            "type": "integer",
            "example": 2500
          },
          "comments": {
            "type": "string",
            "example": "Very good chair for you home"
          },
          "goodtype": {
            "type": "string",
            "enum": [
              "used",
              "new"
            ],
            "example": "new"
          },
          "photo": {
            "type": "string",
            "format": "binary"
          }
        }
      },
      "AddNoticeResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Backend-generated unique identifier",
            "example": "GGHJlhkj668__hjkhjkhh$%"
          },
          "category": {
            "type": "string",
            "description": "notice's category",
            "enum": [
              "auto",
              "business-and-services",
              "for-free",
              "children's-world",
              "home-and-garden",
              "help",
              "electricity",
              "spare-parts-for-transport",
              "fashion-and-style",
              "realty",
              "exchange",
              "repair",
              "work",
              "animals",
              "goods-to-win",
              "hobbies-recreation-sports"
            ]
          },
          "title": {
            "type": "string",
            "description": "the title of user notice",
            "example": "The chair for sale"
          },
          "description": {
            "type": "string",
            "example": "Very good chair for your home"
          },
          "photo": {
            "type": "string",
            "description": "notice image url",
            "example": "photos\\\\2FtTgOgfRKgaEsqQxVhPG_pet.jpg"
          },
          "location": {
            "type": "string",
            "example": "Kyiv"
          },
          "price": {
            "type": "integer",
            "example": 2500
          },
          "comments": {
            "type": "string",
            "description": "additional information about notice or seller",
            "example": "Call from 19:00"
          },
          "createdAt": {
            "type": "string",
            "description": "the creation time of notice (backend-generated)",
            "example": "2023-05-15T10:51:30.531Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "the update time of notice (backend-generated)",
            "example": "2023-05-15T10:51:30.531Z"
          }
        }
      }
    }
  }
}